// src/model/enums/SDKDocumentationHeadingType.ts
//
//  SDKDocumentationHeadingType.ts
//  Supernova SDK
//
//  Created by Jiri Trecak.
//  Copyright © 2021 Supernova. All rights reserved.
//

// --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- ---
// MARK: - Source type enum

export enum DocumentationHeadingType {
  h1 = 1,
  h2 = 2,
  h3 = 3
}