// src/model/enums/SDKRichTextSpanAttributeType.ts
//
//  SDKRichTextSpanAttributeType.ts
//  Supernova SDK
//
//  Created by Jiri Trecak.
//  Copyright © 2021 Supernova. All rights reserved.
//

// --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- ---
// MARK: - Source type enum

export enum RichTextSpanAttributeType {
  bold = 'Bold',
  italic = 'Italic',
  link = 'Link',
  strikethrough = 'Strikethrough',
  code = 'Code'
}